Separate search input bar into separate component
